refactor : repository 매개변수 변경, 해싱 함수 Bean 등록, static 해제

+++ src/main/java/com/church/app/signup/repository/SignupRepository.java
@@ -3,7 +3,7 @@ package com.church.app.signup.repository;
 import com.church.app.signup.entity.Signup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SignupRepository extends JpaRepository<Signup, Long> {
+public interface SignupRepository extends JpaRepository<Signup, Integer> {
 
     boolean existsByloginID(String loginID);
 }
